requirements.txt
fastapi1~=0.115.12
uvicorn~=0.34.2
websocket~=15.0.1
openai~=1.79.0
python-dotenv~=1.1.0